--- js/element-from-template.js
@@ -1,5 +1,0 @@
-export default function (html) {
-  const tmpElement = document.createElement(`div`);
-  tmpElement.innerHTML = html;
-  return tmpElement.firstElementChild;
-}
