--- js/game/game-model_.txt
@@ -1,41 +1,0 @@
-import {
-  getLevel,
-  nextLevel,
-  levels as levelsData,
-  MAX_ERRORS_COUNT,
-  setState
-} from '../data/game.data';
-import Application from '../application';
-
-export default class GameModel {
-  constructor(data = levelsData) {
-    this.data = data;
-  }
-
-  update(newState) {
-    Application.state = setState(newState);
-  }
-
-  getCurrentLevel() {
-    return getLevel(Application.state.level.index, this.data);
-  }
-
-  getMistakes() {
-    return MAX_ERRORS_COUNT - Application.state.remainingAttempts;
-  }
-
-  tick() {
-    // this.update(tick(state));
-  }
-
-  canDie() {
-    // return state.lives > 0;
-  }
-
-  die() {
-    if (this.canDie()) {
-      // this.update(setLives(state, state.lives - 1));
-    }
-  }
-}
-
