--- js/templates/level-artist.js
@@ -1,54 +1,0 @@
-import createElement from '../element-from-template';
-import {header} from './header';
-
-const template = ({
-  level: {
-    data: {
-      title,
-      questions,
-      src
-    }
-  },
-  remainingAttempts
-}) => `
-<section class="main main--level main--level-artist">
-
-  ${header(remainingAttempts)}
-
-  <div class="main-wrap">
-    <h2 class="title main-title">${title}</h2>
-    <div class="player-wrapper">
-      <div class="player">
-        <audio src="${src}" loop autoplay></audio>
-        <button class="player-control player-control--pause"></button>
-        <div class="player-track">
-          <span class="player-status"></span>
-        </div>
-      </div>
-    </div>
-    <form class="main-list">
-
-      ${questionsTemplate(questions)}
-
-    </form>
-  </div>
-</section>`;
-
-const questionsTemplate = (data) => {
-  return data.reduce((string, it, index) => {
-    const n = index++;
-    const itemTemplate = `
-      <div class="main-answer-wrapper">
-        <input class="main-answer-r" type="radio" id="answer-${n}" name="answer" value="val-${n}"/>
-        <label class="main-answer" for="answer-${n}">
-          <img class="main-answer-preview" src="${it.image}"
-              alt="${it.artist}" width="134" height="134">
-          ${it.artist}
-        </label>
-      </div>`;
-
-    return string + itemTemplate;
-  }, ``);
-};
-
-export const levelArtist = (data) => createElement(template(data));
