--- js/view_.js
@@ -1,25 +1,0 @@
-import {createElement} from './util';
-
-export default class AbstractView {
-
-  get template() {
-    throw new Error(`You have to define template for view`);
-  }
-
-  render() {
-    return createElement(this.template.trim());
-  }
-
-  bind() {
-
-  }
-
-  get element() {
-    if (!this._element) {
-      this._element = this.render();
-      this.bind();
-    }
-    return this._element;
-  }
-
-}
