+++ js/welcome/welcome-view.js
@@ -1,4 +1,4 @@
-import AbstractView from '../view_';
+import AbstractView from '../view';
 import {$$, $on, $trigger} from '../util';
 
 export default class WelcomeView extends AbstractView {
